+++ modules/vaos/app/services/vaos/v2/appointments_reason_code_service.rb
@@ -2,12 +2,15 @@
 
 module VAOS
   module V2
-    # Finds the first provider npi id found in provider list and uses that id to request provider name.
+    # This class contains the logic for extracting extra fields from reason codes and updating the appointment.
     class AppointmentsReasonCodeService
       # Modifies the appointment, extracting individual fields from the reason code text whenever possible.
       #
       # @param appointment [Hash] the appointment to modify
       def extract_reason_code_fields(appointment)
+        # Return if the appointment is not a request or is a CC request
+        return if archetype_service.booked?(appointment) || archetype_service.cc?(appointment)
+
         # Retrieve the reason code text, or return if it is not present
         reason_code_text = appointment&.dig(:reason_code, :text)
         return if reason_code_text.nil?
@@ -25,6 +28,10 @@ module VAOS
           contact_info.push({ system: 'email', value: reason_code_hash['email'] })
           appointment[:contact] = { telecom: contact_info }
         end
+      end
+
+      def archetype_service
+        @archetype_service ||= VAOS::V2::AppointmentsArchetypeService.new
       end
     end
   end
