+++ modules/vaos/spec/factories/v2/appointment_forms.rb
@@ -8,9 +8,8 @@ FactoryBot.define do
 
     initialize_with { new(user, attributes) }
 
-    trait :community_cares do
+    trait :community_cares_base do
       kind { 'cc' }
-      status { 'proposed' }
       location_id { '983' }
       service_type { 'podiatry' } # transforms on the front-end need to change
       comment {}
@@ -37,20 +36,6 @@ FactoryBot.define do
           }
         ]
       end
-      contact do
-        {
-          'telecom' => [
-            {
-              'type': 'phone',
-              'value': '2125688887'
-            },
-            {
-              'type': 'email',
-              'value': 'judymorisooooooooooooon@gmail.com'
-            }
-          ]
-        }
-      end
       requested_periods do
         [
           {
@@ -65,6 +50,33 @@ FactoryBot.define do
         {
           'city': 'Helena',
           'state': 'MT'
+        }
+      end
+    end
+
+    trait :community_cares_valid_reason_code_text do
+      community_cares_base
+      status { 'proposed' }
+      reason_code do
+        { 'text': 'station id: 983|preferred modality: FACE TO FACE|phone number: 6195551234|email: myemail72585885@unattended.com|preferred dates:06/26/2024 AM,06/26/2024 PM|reason code:ROUTINEVISIT|comments:test' } # rubocop:disable Layout/LineLength
+      end
+    end
+
+    trait :community_cares do
+      community_cares_base
+      status { 'proposed' }
+      contact do
+        {
+          'telecom' => [
+            {
+              'type': 'phone',
+              'value': '2125688887'
+            },
+            {
+              'type': 'email',
+              'value': 'judymorisooooooooooooon@gmail.com'
+            }
+          ]
         }
       end
       reason_code do
@@ -184,18 +196,35 @@ FactoryBot.define do
       end
     end
 
-    trait :va_booked_invalid_reason_code_text do
-      va_booked_base
+    trait :va_proposed_base do # this has an error, bring up in SOS
+      status { 'proposed' }
+      location_id { '983' }
+      service_type { 'audiology' }
+      comment { 'Follow-up/Routine: testing' }
+      requested_periods do
+        {
+          'end': '2022-01-04T11:59:00Z',
+          'start': '2022-01-04T00:00:00Z'
+        }
+      end
+    end
+
+    trait :va_proposed_valid_reason_code_text do
+      va_proposed_base
+      reason_code do
+        { 'text': 'station id: 983|preferred modality: FACE TO FACE|phone number: 6195551234|email: myemail72585885@unattended.com|preferred dates:06/26/2024 AM,06/26/2024 PM|reason code:ROUTINEVISIT|comments:test' } # rubocop:disable Layout/LineLength
+      end
+    end
+
+    trait :va_proposed_invalid_reason_code_text do
+      va_proposed_base
       reason_code do
         { 'text': 'invalidkeyvaluepair|invalid:key:value:pair||' }
       end
     end
 
     trait :va_proposed do # this has an error, bring up in SOS
-      status { 'proposed' }
-      location_id { '983' }
-      service_type { 'audiology' }
-      comment { 'Follow-up/Routine: testing' }
+      va_proposed_base
       reason_code do
         { 'codeing' => [
             'code': 'Routine Follow-up'
